html/browsers/browsing-the-web/back-forward-cache/test.sub.html

<!doctype html>
<!--
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
-->
<script>
// This Document is opened using `window.open()` with 'noopener' option by
// the main testharness Window (`test2.html`) and
// writes the result back to the main Window via `localStorage`.
// This is because the current test runners expect the top-level testharness
// Document is never unloaded in the middle of a test.

const localStoragePrefix = 'BFCache.test.';

function clearData() {
  const keysToBeRemoved = [];
  for (let i = 0; i < localStorage.length; ++i) {
    const key = localStorage.key(i);
    if (key.startsWith(localStoragePrefix)) {
      keysToBeRemoved.push(key);
    }
  }
  for (const key of keysToBeRemoved) {
    localStorage.removeItem(key);
    console.log(key);
  }
}

function getData(name) {
  return localStorage.getItem(localStoragePrefix + name);
}

function setData(name, value) {
  localStorage.setItem(localStoragePrefix + name, value);
}

function pushData(name, value) {
  const old = getData(name);
  let newArray;
  if (!old) {
    newArray = [value];
  } else {
    newArray = JSON.parse(old);
    newArray.push(value);
  }
  setData(name, JSON.stringify(newArray));
}

function startRecordingEvents() {
  window.testObservedEvents = [];
  let event_list = [
    'visibilitychange',
    'pagehide',
    'pageshow',
    'freeze',
    'resume',
    // Setting unload event handler prevents Firefox's BFCaching.
    // 'unload',
  ];
  for (const event_name of event_list) {
    let result = event_name;
    window.addEventListener(event_name, event => {
      if (event.persisted)
        result += '.persisted';
      pushData('observedEvents', 'window.' + result);
    });
    document.addEventListener(event_name,
        () => pushData('observedEvents', 'document.' + result));
  }
}

// Source: BackForwardCacheBrowserTest.Events

// const t = async_test();

// Initial loading => start of the test.
if (location.search === '') {
  console.log('initialize');
  clearData();
  location.search = '?start';

} else if (getData('state') === null) {
  console.log('start');
  setData('state', 'started');

  window.onload = () => {
    // Navigate after this document is fully loaded.
    // Calling
    // location.href = 'resources/back.html';
    // synchronously here seems to cause `history.back()` on `back.html` to go
    // back to the previous page of this page, not this page, on Firefox.
    setTimeout(() => {
      startRecordingEvents();
      window.addEventListener('pageshow', (() => {
        console.log('pageshow');
        setData('result', getData('observedEvents'));
/*
        assert_equals(getData('state'), 'started');
        assert_array_equals(JSON.parse(getData('observedEvents')), [
          'window.pagehide.persisted',
          'document.visibilitychange',
          'window.visibilitychange',
          // document.freeze/resume are not fired on Firefox.
          // 'document.freeze',
          // 'document.resume',
          'document.visibilitychange',
          'window.visibilitychange',
          'window.pageshow.persisted']);
*/
      }));
      // Cross-site (not only cross-origin) navigation is needed for Chromium.
      // TODO: probably there's an option to enable it for same-site.
      location.href = 'http://{{hosts[alt][www]}}:{{ports[http][0]}}/html/browsers/browsing-the-web/back-forward-cache/resources/back.html';
    }, 100);
  };
} else {
  console.log('unexpected second load');
  setData('result', '["Back-navigated without BFCache"]');
  // t.unreached_func('Back-navigated without BFCache')();
}
</script>
